+++ packages/startupjs/templates/routing/server/index.js
@@ -3,7 +3,6 @@ import startupjsServer from 'startupjs/server'
 import { initApp } from 'startupjs/app/server'
 import orm from '../model'
 import api from './api'
-import getMainRoutes from '../main/routes'
 
 // Init startupjs ORM.
 init({ orm })
@@ -11,10 +10,7 @@ init({ orm })
 // Check '@startupjs/server' readme for the full API
 startupjsServer({
   secure: false, // TODO: https://github.com/startupjs/startupjs#security
-  getHead,
-  appRoutes: [
-    ...getMainRoutes()
-  ]
+  getHead
 }, (ee, options) => {
   initApp(ee)
 
@@ -23,7 +19,7 @@ startupjsServer({
   })
 })
 
-function getHead (appName) {
+function getHead () {
   return `
     <title>App</title>
     <!-- Put vendor JS and CSS here -->