+++ packages/startupjs/templates/ui/server/index.js
@@ -1,11 +1,10 @@
 import init from 'startupjs/init'
 import startupjsServer from 'startupjs/server'
 import { initApp } from 'startupjs/app/server'
-import { initI18n, getI18nRoutes } from 'startupjs/i18n/server'
+import { initI18n } from 'startupjs/i18n/server'
 import { getUiHead, initUi } from '@startupjs/ui/server'
 import orm from '../model'
 import api from './api'
-import getMainRoutes from '../main/routes'
 
 // Init startupjs ORM.
 init({ orm })
@@ -13,11 +12,7 @@ init({ orm })
 // Check '@startupjs/server' readme for the full API
 startupjsServer({
   secure: false, // TODO: https://github.com/startupjs/startupjs#security
-  getHead,
-  appRoutes: [
-    ...getMainRoutes(),
-    ...getI18nRoutes()
-  ]
+  getHead
 }, (ee, options) => {
   initApp(ee)
   initUi(ee, options)
@@ -28,7 +23,7 @@ startupjsServer({
   })
 })
 
-function getHead (appName) {
+function getHead () {
   return `
     ${getUiHead()}
     <title>HelloWorld</title>